Remove ability to append (mad)
This simplifies the code, and works towards being a simple wrapper for the
standard utilities `mv`, `cp` and `ln`. (`cat` is not really the same sort
of thing.)

I freely admit I have never used `mad`, and I can’t think of a sensible use
for it, though given the effort required to implement it, I presume there
was originally a use for it.