Update DIRINFO when creating a directory (fix #24)
When we create a directory, update its DIRINFO. When deciding which file
operation to use to rename, use the live DIRINFO, rather than pre-computing
R_ISX (and remove this flag). Thus, when --makedirs is used, and a
non-existent directory is created, the a subsequent directory move will
succeed, using rename(2) (unless the destination is really on another
device!). Before, the special device code for a non-existent directory would
fool mmv into thinking that a call to copymove was needed, which would
always fail for a directory.